refactor(crm): move companies-by-ids batch endpoint from v2 to v1
The batch endpoint returns only base company details with no v2-specific
transformation, so it belongs alongside the other v1 company lookups
(GET /v1/crm/company/{id}) rather than the v2 store-hydration controller.

Move the endpoint, service method and its integration tests from v2 to v1.
The shared CrmCompanyBatchRequestDto and CrmCompanyDao.findByIdInAndIsDeletedFalse
stay put as they are not v2-specific.
