+++ backend/src/main/java/com/skapp/community/crmplanner/service/CrmCompanyService.java
@@ -3,6 +3,7 @@ package com.skapp.community.crmplanner.service;
 import org.springframework.data.domain.Pageable;
 
 import com.skapp.community.common.payload.response.ResponseEntityDto;
+import com.skapp.community.crmplanner.payload.request.CrmCompanyBatchRequestDto;
 import com.skapp.community.crmplanner.payload.request.CrmCompanyCreateDto;
 import com.skapp.community.crmplanner.payload.request.CrmCompanyDomainSearchRequestDto;
 import com.skapp.community.crmplanner.payload.request.CrmCompanyEditDto;
@@ -21,6 +22,8 @@ public interface CrmCompanyService {
 	ResponseEntityDto getCompanyMetricsById(Long id);
 
 	ResponseEntityDto getCompanyById(Long id);
+
+	ResponseEntityDto getCompaniesByIds(CrmCompanyBatchRequestDto requestDto);
 
 	ResponseEntityDto searchCompaniesByDomain(CrmCompanyDomainSearchRequestDto requestDto);
