+++ backend/src/main/java/com/skapp/community/crmplanner/service/v2/impl/CrmCompanyServiceImplV2.java
@@ -2,10 +2,7 @@ package com.skapp.community.crmplanner.service.v2.impl;
 
 import com.skapp.community.common.payload.response.PageDto;
 import com.skapp.community.common.payload.response.ResponseEntityDto;
-import com.skapp.community.crmplanner.mapper.CrmMapper;
-import com.skapp.community.crmplanner.payload.request.CrmCompanyBatchRequestDto;
 import com.skapp.community.crmplanner.payload.request.CrmCompanyMetricRequestDto;
-import com.skapp.community.crmplanner.payload.response.CrmCompanyResponseDto;
 import com.skapp.community.crmplanner.payload.response.v2.CrmCompanyMetricsResponseDtoV2;
 import com.skapp.community.crmplanner.repository.CrmCompanyDao;
 import com.skapp.community.crmplanner.service.v2.CrmCompanyServiceV2;
@@ -17,17 +14,12 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Collections;
-import java.util.List;
-
 @Service
 @Slf4j
 @RequiredArgsConstructor
 public class CrmCompanyServiceImplV2 implements CrmCompanyServiceV2 {
 
 	private final CrmCompanyDao crmCompanyDao;
-
-	private final CrmMapper crmMapper;
 
 	@Override
 	@Transactional(readOnly = true)
@@ -46,25 +38,6 @@ public class CrmCompanyServiceImplV2 implements CrmCompanyServiceV2 {
 
 		log.info("getCompanyMetrics: execution ended");
 		return new ResponseEntityDto(false, pageDto);
-	}
-
-	@Override
-	@Transactional(readOnly = true)
-	public ResponseEntityDto getCompaniesByIds(CrmCompanyBatchRequestDto requestDto) {
-		log.info("getCompaniesByIds: execution started");
-
-		if (requestDto.getIds() == null || requestDto.getIds().isEmpty()) {
-			log.info("getCompaniesByIds: no ids provided, returning empty list");
-			return new ResponseEntityDto(false, Collections.emptyList());
-		}
-
-		List<CrmCompanyResponseDto> companies = crmCompanyDao.findByIdInAndIsDeletedFalse(requestDto.getIds())
-			.stream()
-			.map(crmMapper::crmCompanyToCrmCompanyResponseDto)
-			.toList();
-
-		log.info("getCompaniesByIds: execution ended with {} result(s)", companies.size());
-		return new ResponseEntityDto(false, companies);
 	}
 
 }
